+++ src/filesage/__main__.py
@@ -1,5 +1,6 @@
-from filesage.cli import main
+from .cli import main
 
 
 if __name__ == "__main__":
     raise SystemExit(main())
+